🍴 Split up the individual formatters into separate files
